1. Refactored BerkeleyDb storage tests into general storage test. 2. Added unit test for in memory storage
